Make sniper rifle more conservative.
The sniper rifle does not need to lock on a target anymore. Instead, pressing down the trigger charges the rifle and slows down the player. When the rifle is charged up, releasing the trigger fires an instant-hit projectile. If the projectile hits an enemy, it causes damage and causes the rifle to fire a homing missile that goes after the same enemy, potentially causing even more damage.